Move structs to go module
